Move get_air_quality_dataset() into air_quality_etl.py
